Add Dockerfile to build and run the application
This commit introduces a Dockerfile that builds a fat JAR of the application using Gradle and then runs it in an OpenJDK 11 container.

The build stage uses a Gradle image to compile the code and create the fat JAR. The run stage uses an OpenJDK 11 image to run the application, exposing port 8080.